requirements.txt.txt
chromadb>=0.4.24
fastapi>=0.109.0
pydantic>=2.5.3
sentence-transformers>=2.2.2
uvicorn>=0.27.0
transformers
sentence-transformers
torch           # required by sentence-transformers
PyPDF2
python-docx

